requirements.txt
azure-mgmt-compute==30.3.0
azure-identity==1.15.0
transformers==4.35.2
skypilot==0.4.1
python-dotenv==0.19.0